Fix keep-alive thread leakage if connect stage was interrupted.
Sometimes an application decides to terminate itself before the
BinaryLogClient successfully connects to a mysql server. The only
possible solution to terminate BinaryLogClient as early as possible is
to attach a LifecycleListeners that calls method `disconnect`.
Unfortunately a keep-alive thread starts after handling lifecycle
listeners, so no one will terminate keep-alive thread, and because it
is not daemonic thread, it prevents application termination.