Fixes for Azure-MySQL connections
Azure has a couple of differences from stock mysql.
First, it always sends a "switch auth" packet back from the first auth
request.  If you google "azure mysql handshake" you'll find a bunch of
other libraries that had to deal with this.

Second, it's rather persnickitty with the response to this; if we send
the response a byte or so at a time azure just closes our connection.