Add master server ID to BinaryLogClient.
When the `BinaryLogClient` connects to a MySQL server, it's not
currently possible to identify to which server it has connected. When
connecting to a pool of replicas through a load balancer, for example, a
reconnect may mean the `BinaryLogClient` is connected to a different
MySQL server.

When using binary log file positions instead of GTIDs, this can be
critical: attempting to resume from the same binary log positions on a
different MySQL server may result in missed or duplicated events.
Similarly, not resuming from the same binary log positions when
reconnecting to the same server is wasteful and may result in longer
recovery/catch-up times.

This change adds the ability to get the equivalent of `MASTER_SERVER_ID`
from the `BinaryLogClient`. This can then be inspected within each of
the `LifecycleListener` events.